+++ IIFYM_Main/src/main/res/values/dimens.xml
@@ -15,4 +15,8 @@
     <dimen name="labels_text_size">15sp</dimen>
     <dimen name="fab_size_normal">56dp</dimen>
     <dimen name="fab_size_mini">40dp</dimen>
+
+    <dimen name="tab_max_width">264dp</dimen>
+    <dimen name="tab_padding_bottom">16dp</dimen>
+    <dimen name="tab_label">14sp</dimen>
 </resources>
